commands related to setting public link's expiry date is moved to publicLinksDatePicker
