+++ tests/acceptance/pageObjects/FilesPageElement/publicLinksDialog.js
@@ -24,95 +24,6 @@ module.exports = {
         .waitForOutstandingAjaxCalls()
     },
     /**
-     * sets up the xpath for year in expiry date of public link
-     *
-     * @param year
-     * @returns {{locateStrategy: string, selector: *}}
-     */
-    setExpiryDateYearSelectorXpath: function (year) {
-      const yearSelectorXpath = util.format(this.elements.dateTimeYearPicker.selector, year)
-      return {
-        selector: yearSelectorXpath,
-        locateStrategy: this.elements.dateTimeYearPicker.locateStrategy
-      }
-    },
-    /**
-     * sets up the xpath for month in expiry date of public link
-     *
-     * @param month
-     * @returns {{locateStrategy: string, selector: *}}
-     */
-    setExpiryDateMonthSelectorXpath: function (month) {
-      const monthSelectorXpath = util.format(this.elements.dateTimeMonthPicker.selector, month)
-      return {
-        selector: monthSelectorXpath,
-        locateStrategy: this.elements.dateTimeMonthPicker.locateStrategy
-      }
-    },
-    /**
-     * sets up the xpath for year in expiry date of public link
-     *
-     * @param day
-     * @returns {{locateStrategy: string, selector: *}}
-     */
-    setExpiryDateDaySelectorXpath: function (day) {
-      const daySelectorXpath = util.format(this.elements.dateTimeDayPicker.selector, day)
-      return {
-        selector: daySelectorXpath,
-        locateStrategy: this.elements.dateTimeDayPicker.locateStrategy
-      }
-    },
-    /**
-     * sets provided year in expiry date field on webUI
-     *
-     * @param {string} year
-     * @returns {Promise<void>}
-     */
-    setExpiryDateYear: function (year) {
-      const yearSelector = this.setExpiryDateYearSelectorXpath(year)
-      return this
-        .waitForElementVisible('@dateTimePopupYear')
-        .waitForAnimationToFinish()
-        .waitForElementEnabled(
-          this.elements.dateTimePopupYear.selector
-        )
-        .click('@dateTimePopupYear')
-        .waitForElementVisible(yearSelector)
-        .click(yearSelector)
-        .click('@dateTimeOkButton')
-        .waitForElementNotPresent(yearSelector)
-    },
-    /**
-     * sets provided month in expiry date field on webUI
-     *
-     * @param {string} month
-     * @returns {Promise<void>}
-     */
-    setExpiryDateMonth: function (month) {
-      const monthSelector = this.setExpiryDateMonthSelectorXpath(month)
-      return this
-        .waitForElementVisible('@dateTimePopupDate')
-        .click('@dateTimePopupDate')
-        .waitForElementVisible(monthSelector)
-        .click(monthSelector)
-        .click('@dateTimeOkButton')
-        .waitForElementNotPresent(monthSelector)
-    },
-    /**
-     * sets provided day in expiry date field on webUI
-     *
-     * @param {string} day
-     * @returns {Promise<void>}
-     */
-    setExpiryDateDay: function (day) {
-      const daySelector = this.setExpiryDateDaySelectorXpath(day)
-      return this
-        .waitForElementVisible(daySelector)
-        .click(daySelector)
-        .click('@dateTimeOkButton')
-        .waitForElementNotPresent(daySelector)
-    },
-    /**
      * sets role or permissions for public link on webUI
      *
      * @param {string} role - e.g. Viewer, Contributor, Editor, Uploader
@@ -160,7 +71,8 @@ module.exports = {
      * @param {string} value - provided date in format YYYY-MM-DD, or empty string to unset date
      * @returns {Promise}
      */
-    setPublicLinkExpiryDate: function (value) {
+    setPublicLinkExpiryDate: async function (value) {
+      const publicLinkDatePicker = this.api.page.FilesPageElement.publicLinksDatePicker()
       if (value === '') {
         return this.click('@publicLinkDeleteExpirationDateButton')
       }
@@ -170,10 +82,11 @@ module.exports = {
       const month = dateToSet.toLocaleString('en-GB', { month: 'long' })
       const day = dateToSet.getDate()
 
-      return this
+      await this
         .initAjaxCounters()
         .waitForElementVisible('@linkExpirationDateField')
         .click('@linkExpirationDateField')
+      return publicLinkDatePicker
         .setExpiryDateYear(year)
         .setExpiryDateMonth(month)
         .setExpiryDateDay(day)
@@ -196,63 +109,6 @@ module.exports = {
         return this.setPublicLinkExpiryDate(value)
       }
       return this
-    },
-    /**
-     * checks if the given expiryDate is disabled or not
-     *
-     * @param {string} linkName name of the public link
-     * @param {string} pastDate provided past date for inspection
-     *  pastDate should be in form 2000-August-7 | 2000-Aug-7
-     *  leading zeros before day are removed
-     * @returns {Promise<boolean>}
-     */
-    isExpiryDateDisabled: async function (linkName, pastDate) {
-      await this.clickLinkEditBtn(linkName)
-      const [year, month, day] = pastDate.split(/-/)
-      let disabled = false
-      const iDay = parseInt(day)
-      const yearSelector = this.setExpiryDateYearSelectorXpath(year)
-      const monthSelector = this.setExpiryDateMonthSelectorXpath(month)
-      const daySelector = this.setExpiryDateDaySelectorXpath(iDay)
-      await this
-        .initAjaxCounters()
-        .waitForElementVisible('@linkExpirationDateField')
-        .click('@linkExpirationDateField')
-        .waitForElementVisible('@dateTimePopupYear')
-        .waitForAnimationToFinish()
-        .waitForElementEnabled(
-          this.elements.dateTimePopupYear.selector
-        )
-        .angryClick('@dateTimePopupYear')
-        .waitForElementVisible(yearSelector)
-        .getAttribute(yearSelector, 'class', (result) => {
-          if (result.value.includes('--disabled') === true) {
-            disabled = true
-          }
-        })
-      if (disabled) { return disabled }
-      await this
-        .click(yearSelector)
-        .click('@dateTimeOkButton')
-        .waitForElementVisible('@dateTimePopupDate')
-        .click('@dateTimePopupDate')
-        .waitForElementVisible(monthSelector)
-        .getAttribute(monthSelector, 'class', (result) => {
-          if (result.value.includes('--disabled') === true) {
-            disabled = true
-          }
-        })
-      if (disabled) { return disabled }
-      await this
-        .click(monthSelector)
-        .click('@dateTimeOkButton')
-        .waitForElementVisible(daySelector)
-        .getAttribute(daySelector, 'class', (result) => {
-          if (result.value.includes('--disabled') === true) {
-            disabled = true
-          }
-        })
-      return disabled
     },
     /**
      * sets up public link share edit form
@@ -416,9 +272,15 @@ module.exports = {
      */
     getPublicLinkUrls: async function () {
       const promiseList = []
-      const publicLinkUrlXpath = this.elements.publicLinkContainer.selector + this.elements.publicLinkInformation.selector + this.elements.publicLinkUrl.selector
+      const publicLinkUrlXpath = this.elements.publicLinkContainer.selector +
+        this.elements.publicLinkInformation.selector +
+        this.elements.publicLinkUrl.selector
       await this.initAjaxCounters()
-        .waitForElementPresent({ locateStrategy: 'xpath', selector: publicLinkUrlXpath, abortOnFailure: false })
+        .waitForElementPresent({
+          locateStrategy: 'xpath',
+          selector: publicLinkUrlXpath,
+          abortOnFailure: false
+        })
         .waitForOutstandingAjaxCalls()
         .api.elements('xpath', publicLinkUrlXpath, result => {
           result.value.map(item => {
@@ -563,36 +425,8 @@ module.exports = {
     publicLinkSaveButton: {
       selector: '#oc-files-file-link-save'
     },
-    publicLinkDeleteExpirationDateButton: {
-      selector: '#oc-files-file-link-expire-date-delete'
-    },
-    linkExpirationDateField: {
-      selector: '.vdatetime-input'
-    },
     dateTimePopup: {
       selector: '.vdatetime-popup'
-    },
-    dateTimePopupYear: {
-      selector: '.vdatetime-popup__year'
-    },
-    dateTimePopupDate: {
-      selector: '.vdatetime-popup__date'
-    },
-    dateTimeMonthPicker: {
-      selector: '//div[@class="vdatetime-month-picker"]//div[contains(text(),"%s")]',
-      locateStrategy: 'xpath'
-    },
-    dateTimeYearPicker: {
-      selector: '//div[@class="vdatetime-year-picker"]//div[normalize-space(.)="%s"]',
-      locateStrategy: 'xpath'
-    },
-    dateTimeDayPicker: {
-      selector: '//div[@class="vdatetime-calendar"]//span/span[normalize-space(.)="%s"]/../..',
-      locateStrategy: 'xpath'
-    },
-    dateTimeOkButton: {
-      selector: '//div[@class="vdatetime-popup__actions"]/div[.="Ok"]',
-      locateStrategy: 'xpath'
     },
     dateTimeCancelButton: {
       selector: '//div[@class="vdatetime-popup__actions"]/div[.="Cancel"]',
@@ -603,6 +437,12 @@ module.exports = {
     },
     sidebarPrivateLinkIconCopied: {
       selector: '#files-sidebar-private-link-icon-copied'
+    },
+    publicLinkDeleteExpirationDateButton: {
+      selector: '#oc-files-file-link-expire-date-delete'
+    },
+    linkExpirationDateField: {
+      selector: '.vdatetime-input'
     }
   }
 }
